First pass at refactoring tests

+++ dirstuff/os/filesystem.py
@@ -1,17 +1,20 @@
 import re
 import shutil
 from pathlib import Path as PathlibPath
-from typing import Any, Iterator, Protocol
+from typing import Any, Iterator, Union
 
 
-class Path(Protocol):
+class Path:
     """Abstract base class for system paths."""
 
     libpath: PathlibPath
 
-    def __init__(self, *args: Any, **kwargs: Any):
+    def __init__(self, path: Union[str, PathlibPath, "Path"], *args: Any, **kwargs: Any):
         """Construct a Path object."""
-        self.libpath = PathlibPath(*args, **kwargs)
+        if isinstance(path, Path):
+            self.libpath = path.libpath
+        else:
+            self.libpath = PathlibPath(path, *args, **kwargs)
 
     @property
     def name(self) -> str:
@@ -26,6 +29,15 @@ class Path(Protocol):
         """Get the string representation of the path."""
         return str(self.libpath)
 
+    @property
+    def parent(self) -> "Path":
+        """Get the parent path.
+
+        Returns:
+            Dir: The parent path.
+        """
+        return Path(self.libpath.parent)
+
 
 class Dir(Path):
     """A directory."""
@@ -33,13 +45,13 @@ class Dir(Path):
     def __init__(self, *args: Any, **kwargs: Any):
         """Construct a Dir object.
 
-        This constructor takes the same arguments as the pathlib.Path constructor.
+        This constructor takes the same arguments as the pathlib.Path constructor or an instance of dirstuff.Path.
 
         Raises:
             ValueError: If the path is not a directory.
         """
         super().__init__(*args, **kwargs)
-        if self.libpath.exists() and not self.libpath.is_dir():
+        if self.exists() and not self.libpath.is_dir():
             msg = f"Tried to create Dir from non-dir path: {self.libpath}"
             raise ValueError(msg)
 
@@ -101,7 +113,7 @@ class Dir(Path):
         if new_name == self.name and not same_name_ok:
             msg = f"No change made to dir name: {self.name}"
             raise ValueError(msg)
-        if not self.libpath.exists():
+        if not self.exists():
             msg = f"Dir does not exist: {self.libpath}"
             raise FileNotFoundError(msg)
         new_path = self.libpath.parent / new_name
@@ -127,7 +139,7 @@ class Dir(Path):
         if new_name == self.name and not same_name_ok:
             msg = f"No change made to dir name: {self.name}"
             raise ValueError(msg)
-        if not self.libpath.exists():
+        if not self.exists():
             msg = f"Dir does not exist: {self.libpath}"
             raise FileNotFoundError(msg)
         new_path = self.libpath.parent / new_name
@@ -142,23 +154,38 @@ class Dir(Path):
 
         Returns:
             Dir: The moved directory.
+
+        Raises:
+            FileNotFoundError: If the destination directory does not exist.
         """
-        path = dir.libpath / self.libpath.name
+        if not dir.exists():
+            msg = f"Destination directory does not exist: {dir.libpath}"
+            raise FileNotFoundError(msg)
+        path = dir.libpath / self.name
         shutil.move(self.libpath, path)
         self.libpath = PathlibPath(path)
         return self
 
-    def copy_to(self, dir: "Dir") -> "Dir":
+    def copy_to(self, path: Path, overwrite_ok: bool = False) -> "Dir":
         """Copy the directory to another directory.
 
         Args:
-            dir (Dir): The directory to copy to.
+            path (Path): The path to copy to.
+            overwrite_ok (bool): Whether to allow overwriting the destination directory. Defaults to False.
 
         Returns:
             Dir: The copied directory.
+
+        Raises:
+            FileExistsError: If the destination directory already exists and overwrite_ok is False.
         """
-        shutil.copytree(self.libpath, dir.libpath)
-        self.libpath = dir.libpath
+        if path.exists():
+            if not overwrite_ok:
+                msg = f"Destination path already exists: {path.libpath}"
+                raise FileExistsError(msg)
+            shutil.rmtree(path.libpath)
+        shutil.copytree(self.libpath, path.libpath)
+        self.libpath = path.libpath
         return self
 
     def copy_into(self, dir: "Dir") -> "Dir":
@@ -170,7 +197,10 @@ class Dir(Path):
         Returns:
             Dir: The copied directory
         """
-        path = dir.libpath / self.libpath.name
+        if not dir.exists():
+            msg = f"Destination directory does not exist: {dir.libpath}"
+            raise FileNotFoundError(msg)
+        path = dir.libpath / self.name
         shutil.copytree(self.libpath, path)
         self.libpath = path
         return self
@@ -184,7 +214,7 @@ class Dir(Path):
         Raises:
             FileNotFoundError: If the directory does not exist and missing_ok is False.
         """
-        if not self.libpath.exists() and not missing_ok:
+        if not self.exists() and not missing_ok:
             msg = f"No directory at {self.libpath}"
             raise FileNotFoundError(msg)
         shutil.rmtree(self.libpath)
@@ -208,8 +238,23 @@ class Dir(Path):
 
         Returns:
             Dir: The parent directory.
+
+        Raises:
+            FileNotFoundError: If the parent directory does not exist.
         """
+        if not self.exists():
+            msg = f"Dir does not exist: {self.libpath}"
+            raise FileNotFoundError(msg)
         return Dir(self.libpath.parent)
+
+    @property
+    def path(self) -> Path:
+        """Get the path for the dir.
+
+        Returns:
+            Path: The path for the dir.
+        """
+        return Path(self.libpath)
 
 
 class File(Path):
@@ -218,13 +263,13 @@ class File(Path):
     def __init__(self, *args: Any, **kwargs: Any):
         """Construct a File object.
 
-        This constructor takes the same arguments as the pathlib.Path constructor.
+        This constructor takes the same arguments as the pathlib.Path constructor or an instance of dirstuff.Path.
 
         Raises:
             ValueError: If the path is not a file.
         """
         super().__init__(*args, **kwargs)
-        if self.libpath.exists() and not self.libpath.is_file():
+        if self.exists() and not self.libpath.is_file():
             msg = f"Tried to create File from non-file path: {self.libpath}"
             raise ValueError(msg)
 
@@ -245,7 +290,7 @@ class File(Path):
         if new_name == self.name and not same_name_ok:
             msg = f"No change made to file name: {self.name}"
             raise ValueError(msg)
-        if not self.libpath.exists():
+        if not self.exists():
             msg = f"File does not exist: {self.libpath}"
             raise FileNotFoundError(msg)
         new_path = self.libpath.parent / new_name
@@ -272,13 +317,13 @@ class File(Path):
             msg = f"No change made to file name: {self.name}"
             raise ValueError(msg)
         new_path = self.libpath.parent / new_name
-        if not self.libpath.exists():
+        if not self.exists():
             msg = f"File does not exist: {self.libpath}"
             raise FileNotFoundError(msg)
         self.libpath = self.libpath.rename(new_path)
         return self
 
-    def move_into(self, dir: "Dir") -> "File":
+    def move_into(self, dir: Dir) -> "File":
         """Move the file into a directory.
 
         Args:
@@ -286,25 +331,40 @@ class File(Path):
 
         Returns:
             File: The moved file.
+
+        Raises:
+            FileNotFoundError: If the destination directory does not exist.
         """
-        path = dir.libpath / self.libpath.name
+        if not dir.exists():
+            msg = f"Destination directory does not exist: {dir.libpath}"
+            raise FileNotFoundError(msg)
+        path = dir.libpath / self.name
         self.libpath = self.libpath.rename(path)
         return self
 
-    def copy_to(self, file: "File") -> "File":
-        """Copy the file to another file.
+    def copy_to(self, path: Path, overwrite_ok: bool = False) -> "File":
+        """Copy the file to another path.
 
         Args:
-            file (File): The file to copy to.
+            path (Path): The path to copy to.
+            overwrite_ok (bool): Whether to allow overwriting the destination directory. Defaults to False.
 
         Returns:
             File: The copied file.
+
+        Raises:
+            FileExistsError: If the destination directory already exists and overwrite_ok is False.
         """
-        shutil.copy(self.libpath, file.libpath)
-        self.libpath = file.libpath
+        if path.exists():
+            if not overwrite_ok:
+                msg = f"Destination path already exists: {path.libpath}"
+                raise FileExistsError(msg)
+            path.libpath.unlink(missing_ok=True)
+        shutil.copy(self.libpath, path.libpath)
+        self.libpath = path.libpath
         return self
 
-    def copy_into(self, dir: "Dir") -> "File":
+    def copy_into(self, dir: Dir) -> "File":
         """Copy the file into a directory.
 
         Args:
@@ -312,8 +372,14 @@ class File(Path):
 
         Returns:
             File: The copied file.
+
+        Raises:
+            FileNotFoundError: If the destination directory does not exist.
         """
-        path = dir.libpath / self.libpath.name
+        if not dir.exists():
+            msg = f"Destination directory does not exist: {dir.libpath}"
+            raise FileNotFoundError(msg)
+        path = dir.libpath / self.name
         shutil.copy(self.libpath, path)
         self.libpath = path
         return self
@@ -324,6 +390,9 @@ class File(Path):
         Args:
             missing_ok (bool): Whether to allow the file to be missing. Defaults to False.
         """
+        if not self.exists() and not missing_ok:
+            msg = f"File does not exist: {self.libpath}"
+            raise FileNotFoundError(msg)
         self.libpath.unlink(missing_ok=missing_ok)
 
     @property
@@ -345,10 +414,25 @@ class File(Path):
         return self.libpath.suffix
 
     @property
-    def parent(self) -> "Dir":
+    def parent(self) -> Dir:
         """Get the parent directory.
 
         Returns:
             Dir: The parent directory.
+
+        Raises:
+            FileNotFoundError: If the parent directory does not exist.
         """
+        if not self.exists():
+            msg = f"File does not exist: {self.libpath}"
+            raise FileNotFoundError(msg)
         return Dir(self.libpath.parent)
+
+    @property
+    def path(self) -> Path:
+        """Get the path for the file.
+
+        Returns:
+            Path: The path for the file.
+        """
+        return Path(self.libpath)
